+++ tests/conftest.py
@@ -12,6 +12,7 @@ import pandas as pd
 import polars as pl
 import pytest
 from anndata import AnnData
+
 from pixelator.config import AntibodyPanel
 from pixelator.graph import update_edgelist_membership
 from pixelator.graph.utils import union as graph_union
@@ -20,7 +21,6 @@ from pixelator.pixeldataset import (
 )
 from pixelator.pixeldataset.precomputed_layouts import PreComputedLayouts
 from pixelator.pixeldataset.utils import edgelist_to_anndata
-
 from tests.graph.networkx.test_tools import (
     create_fully_connected_bipartite_graph,
     create_random_graph,