+++ tests/report/test_data_collection.py
@@ -6,6 +6,7 @@ Copyright © 2023 Pixelgen Technologies AB.
 import json
 
 import pytest
+
 from pixelator.report.common.json_encoder import PixelatorJSONEncoder
 from pixelator.report.qcreport.collect import (
     collect_antibody_counts_data,