+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.3] - 2024-09-26
+
+### Fixed
+
+ - Improved memory usage when aggregating PXL files with precomputed layouts.
+
 ## [0.18.2] - 2024-07-16
 
 ### Changed